The media tray moves behind useComposerMedia
Pick, edit, upload, remove — the one part of a draft that persists without a
save. It keys on the route id exactly as before, so the tray still works from
the moment the URL has one.
